Moved the common test case functionality to a trait.
With this change, we can now reuse the functionality in any test
suite without duplication.
