+++ tests/phpunit/test-case-trait.php
@@ -1,0 +1,118 @@
+<?php
+/**
+ * Common functionality for the Beans Test Suites.
+ *
+ * @package Beans\Framework\Tests
+ *
+ * @since   1.5.0
+ */
+
+namespace Beans\Framework\Tests;
+
+trait Test_Case_Trait {
+
+	/**
+	 * Load the original Beans' functions into memory before we start.
+	 *
+	 * Then in our tests, we monkey patch via Brain Monkey, which redefines the original function.
+	 * At tear down, the original function is restored in Brain Monkey, by calling Patchwork\restoreAll().
+	 *
+	 * @since 1.5.0
+	 *
+	 * @param array $files Array of files to load into memory.
+	 *
+	 * @return void
+	 */
+	protected function load_original_functions( array $files ) {
+
+		foreach ( $files as $file ) {
+			require_once BEANS_TESTS_LIB_DIR . $file;
+		}
+	}
+
+	/**
+	 * Format the HTML by stripping out the whitespace between the HTML tags and then putting each tag on a separate
+	 * line.
+	 *
+	 * Why? We can then compare the actual vs. expected HTML patterns without worrying about tabs, new lines, and extra
+	 * spaces.
+	 *
+	 * @since 1.5.0
+	 *
+	 * @param string $html HTML to strip.
+	 *
+	 * @return string
+	 */
+	protected function format_the_html( $html ) {
+		$html = trim( $html );
+
+		// Strip whitespace between the tags.
+		$html = preg_replace( '/(\>)\s*(\<)/m', '$1$2', $html );
+
+		// Strip whitespace at the end of a tag.
+		$html = preg_replace( '/(\>)\s*/m', '$1$2', $html );
+
+		// Strip whitespace at the start of a tag.
+		$html = preg_replace( '/\s*(\<)/m', '$1$2', $html );
+
+		return str_replace( '>', ">\n", $html );
+	}
+
+	/**
+	 * Get reflective access to the private method.
+	 *
+	 * @since 1.5.0
+	 *
+	 * @param string $method_name Method name for which to gain access.
+	 * @param string $class_name  Name of the target class.
+	 *
+	 * @return \ReflectionMethod
+	 * @throws \ReflectionException Throws an exception if method does not exist.
+	 */
+	protected function get_reflective_method( $method_name, $class_name ) {
+		$class  = new \ReflectionClass( $class_name );
+		$method = $class->getMethod( $method_name );
+		$method->setAccessible( true );
+
+		return $method;
+	}
+
+	/**
+	 * Get reflective access to the private property.
+	 *
+	 * @since 1.5.0
+	 *
+	 * @param string       $property Property name for which to gain access.
+	 * @param string|mixed $class    Class name or instance.
+	 *
+	 * @return \ReflectionProperty|string
+	 * @throws \ReflectionException Throws an exception if property does not exist.
+	 */
+	protected function get_reflective_property( $property, $class ) {
+		$class    = new \ReflectionClass( $class );
+		$property = $class->getProperty( $property );
+		$property->setAccessible( true );
+
+		return $property;
+	}
+
+	/**
+	 * Set the value of a property or private property.
+	 *
+	 * @since 1.5.0
+	 *
+	 * @param mixed  $value    The value to set for the property.
+	 * @param string $property Property name for which to gain access.
+	 * @param mixed  $instance Instance of the target object.
+	 *
+	 * @return \ReflectionProperty|string
+	 * @throws \ReflectionException Throws an exception if property does not exist.
+	 */
+	protected function set_reflective_property( $value, $property, $instance ) {
+		$property = $this->get_reflective_property( $property, $instance );
+		$property->setValue( $instance, $value );
+		$property->setAccessible( false );
+
+		return $property;
+	}
+}
